+++ feature/auth/src/main/java/com/anna/greeneats/auth/login/state/LoginScreenViewModel.kt
@@ -7,14 +7,13 @@ import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.anna.greeneats.auth.common.EmailErrorState
 import com.anna.greeneats.auth.common.PasswordErrorState
-import com.anna.greeneats.core.model.resource.Resource
 import com.anna.greeneats.core.util.validation.error.Email
 import com.anna.greeneats.core.util.validation.error.Password
 import com.anna.greeneats.core.util.validation.main.EmailValidations
 import com.anna.greeneats.core.util.validation.main.PasswordValidations
 import com.anna.greeneats.data.auth.AuthRepository
+import com.anna.greeneats.model.main.Resource
 import dagger.hilt.android.lifecycle.HiltViewModel
-import dagger.hilt.android.qualifiers.ActivityContext
 import kotlinx.coroutines.launch
 import javax.inject.Inject
